WIP: Add behavioral tests
