+++ Package.swift
@@ -12,6 +12,7 @@ let package = Package(
         .package(url: "https://github.com/eneko/MarkdownGenerator.git", from: "0.4.0"),
         .package(url: "https://github.com/onevcat/Rainbow", from: "3.0.0"),
         .package(url: "https://github.com/thoughtbot/Curry.git", from: "4.0.1"),
+        .package(url: "https://github.com/eneko/System.git", from: "0.1.0")
     ],
     targets: [
         .target(name: "SourceDocs", dependencies: [
@@ -21,7 +22,8 @@ let package = Package(
             "Commandant",
             "Curry"
             ]),
-        .testTarget(name: "SourceDocsTests", dependencies: ["SourceDocs"]),
+        .testTarget(name: "BehavioralTests", dependencies: ["System"]),
+        .testTarget(name: "UnitTests", dependencies: ["SourceDocs"]),
         .target(name: "SourceDocsDemo", dependencies: []),
     ]
 )
