+++ Tests/UnitTests/AccessLevelTests.swift
@@ -1,0 +1,18 @@
+//
+//  AccessLevelTests.swift
+//  
+//
+//  Created by Eneko Alonso on 11/1/19.
+//
+
+import XCTest
+// TODO: import SourceDocs
+
+class AccessLevelTests: XCTestCase {
+
+    func testPriority() {
+//        let levels = AccessLevel.allCases
+//        XCTAssertEqual(levels.count, 5)
+    }
+
+}
